+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.1.0 (TBD)
 
 - Annotated a number of non-final classes as `@final`, they will be made final in 3.0
+- [#21](https://github.com/GeniusesOfSymfony/PubSubRouterBundle/issues/21) - Ensure a RouteCollection is returned from the router if there are no resources
 
 ## 2.0.0 (2020-01-08)